Component-ified Intro Text
