+++ .eslintrc.js
@@ -19,5 +19,6 @@ module.exports = {
   ],
   // add your custom rules here
   rules: {
+    'vue/attribute-hyphenation': 'off'
   }
 }
