+++ pages/cms.vue
@@ -11,51 +11,10 @@
         {{ w }}
       </button>
     </div>
-    <div id="intro">
-      <div id="intro-header">
-        <p>Intro Text</p>
-        <button v-if="!editingIntro" @click="startEditingIntro">Edit</button>
-        <p v-if="editingIntro" id="intro-cancel" @click="stopEditingIntro">
-          Cancel
-        </p>
-        <p v-if="editingIntro" id="intro-save" @click="saveEditingIntro">
-          Save
-        </p>
-      </div>
-      <div v-if="!editingIntro" class="intro-body">
-        <p>{{ introTexts[selectedWebsite].introText }}</p>
-        <p>{{ introTexts[selectedWebsite].introSubtext }}</p>
-      </div>
-      <div v-if="editingIntro" class="intro-body">
-        <p
-          v-if="
-            introTexts[selectedWebsite].introLastEditedBy &&
-              introTexts[selectedWebsite].introLastEditedBy
-          "
-          class="last-edited-by"
-        >
-          Last Edited By:
-          {{ introTexts[selectedWebsite].introLastEditedBy }} at
-          {{
-            new Date(
-              introTexts[selectedWebsite].introLastEditedDate
-            ).toISOString()
-          }}
-        </p>
-        <div id="title">
-          <p>Title</p>
-        </div>
-        <textarea
-          v-model="introTexts[selectedWebsite].introText"
-          class="intro-input"
-        />
-        <p>Description</p>
-        <textarea
-          v-model="introTexts[selectedWebsite].introSubtext"
-          class="intro-input"
-        />
-      </div>
-    </div>
+    <IntroText
+      :originalIntroTexts="introTexts"
+      :selectedWebsite="selectedWebsite"
+    ></IntroText>
     <div id="files-select">
       <div class="large-12 medium-12 small-12 cell">
         <label id="files-label">Images</label>
@@ -87,7 +46,7 @@
     </div>
     <Faq
       :website="selectedWebsite"
-      :listOfFaq="faq"
+      :list-of-faq="faq"
       @refreshData="refreshData(selectedWebsite)"
     />
   </div>
@@ -97,12 +56,14 @@
 /* eslint-disable no-console,import/no-duplicates,prettier/prettier */
 
 import firebase from 'firebase/app'
-import { auth } from '../plugins/firebase'
+import { auth } from '~/plugins/firebase'
 import fireDb from '~/plugins/firebase'
 import Faq from '~/components/FAQ.vue'
+import IntroText from '../components/IntroText'
 
 export default {
   components: {
+    IntroText,
     Faq
   },
   async asyncData({ redirect }) {
@@ -119,30 +80,10 @@ export default {
       introTexts: introTexts,
       selectedWebsite: selectedWebsite,
       files: [],
-      faq: [],
-      editingIntro: false
+      faq: []
     }
   },
   methods: {
-    startEditingIntro() {
-      this.editingIntro = true
-    },
-    stopEditingIntro() {
-      this.editingIntro = false
-    },
-    async saveEditingIntro() {
-      this.introTexts[this.selectedWebsite].introLastEditedBy = firebase.auth().currentUser.email
-      this.introTexts[this.selectedWebsite].introLastEditedDate = Date.now()
-
-      await fireDb.updateIntroText(
-        this.selectedWebsite,
-        this.introTexts[this.selectedWebsite].introText,
-        this.introTexts[this.selectedWebsite].introSubtext,
-        this.introTexts[this.selectedWebsite].introLastEditedBy,
-        this.introTexts[this.selectedWebsite].introLastEditedDate
-      )
-      this.stopEditingIntro()
-    },
     async changeWebsite(e) {
       this.selectedWebsite = e.target.value
       await this.refreshData()
